Uniformly apply Windows logic in cpp_extensions everywhere
Pull Request resolved: https://github.com/pytorch/pytorch/pull/31161


Previously, it wasn't necessary to specify `DT_NEEDED` in C++ extensions on Linux (aka pass `-l` flags) because all of the symbols would have already been loaded with `RTLD_GLOBAL`, so there wouldn't be any undefined symbols.  But when we switch to loading `_C` with `RTLD_LOCAL`, it's now necessary for all the C++ extensions to know what libraries to link with. The resulting code is clearer and more uniform, so it's wins all around.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

Differential Revision: [D19262578](https://our.internmc.facebook.com/intern/diff/D19262578/)
ghstack-source-id: 96364920